Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
google-generativeai==0.3.2
shap==0.44.0
lime==0.2.0.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.6 